All color related values converted to bytes

+++ drivers/chainable-rgbled-grove/chainable-rgbled-grove/RGB.cs
@@ -5,11 +5,11 @@ namespace chainable_rgbled_grove
 {
     public class RGB
     {
-        public int red;
-        public int green;
-        public int blue;
+        public byte red;
+        public byte green;
+        public byte blue;
 
-        public RGB(int red, int green, int blue)
+        public RGB(byte red, byte green, byte blue)
         {
             this.red = red;
             this.green = green;
